Rename package and class to comply with coding standard
