+++ src/main/java/universe/MissingListItemException.java
@@ -1,11 +1,11 @@
-package UNIverse;
+package universe;
 
 /**
- * MissingListItemException class inherits from UNIverseException.
+ * MissingListItemException class inherits from UniverseException.
  * It passes on the error message that the chore does not exist.
  */
 
-public class MissingListItemException extends UNIverseException {
+public class MissingListItemException extends UniverseException {
     public MissingListItemException() {
         super("This chore is not in the list :(");
     }
