+++ src/main/java/universe/MissingStartAndEndException.java
@@ -1,10 +1,10 @@
-package UNIverse;
+package universe;
 
 /**
- * MissingStartAndEndException class inherits from UNIverseException.
+ * MissingStartAndEndException class inherits from UniverseException.
  * It passes on the error message that the event is lacking either a start or end time or both.
  */
-public class MissingStartAndEndException extends UNIverseException {
+public class MissingStartAndEndException extends UniverseException {
     public MissingStartAndEndException() {
         super("An event must have a start and end time!");
     }
