+++ src/main/java/universe/WaitForResponse.java
@@ -1,4 +1,4 @@
-package UNIverse;
+package universe;
 
 import java.util.Scanner;
