+++ text-ui-test/EXPECTED.TXT
@@ -1,34 +1,25 @@
-Welcome to the UNIverse!
+Welcome to the Universe!
 What cosmic quest can I assist you with today?
-
 [T][ ] read psych chap 1 added to Checklist!
 You now have 1 chore(s) in the checklist.
-
 [D][ ] cs2103t ip week 2 (by: 24 Jan 2025) added to Checklist!
 You now have 2 chore(s) in the checklist.
-
 [E][ ] ihg track (from: 6 Feb 2025 7pm to: 9pm) added to Checklist!
 You now have 3 chore(s) in the checklist.
-
 Cosmic Chore Checklist:
 1 [T][ ] read psych chap 1
 2 [D][ ] cs2103t ip week 2 (by: 24 Jan 2025)
-3 [E][ ] ihg track (from: 5 Feb 2025 7pm to: 9pm)
-
+3 [E][ ] ihg track (from: 6 Feb 2025 7pm to: 9pm)
 Yay read psych chap 1 successfully completed!
 [T][X] read psych chap 1
-
 Cosmic Chore Checklist:
 1 [T][X] read psych chap 1
 2 [D][ ] cs2103t ip week 2 (by: 24 Jan 2025)
 3 [E][ ] ihg track (from: 6 Feb 2025 7pm to: 9pm)
-
 read psych chap 1 marked as not done yet :(
 [T][ ] read psych chap 1
-
 Cosmic Chore Checklist:
 1 [T][ ] read psych chap 1
 2 [D][ ] cs2103t ip week 2 (by: 24 Jan 2025)
 3 [E][ ] ihg track (from: 6 Feb 2025 7pm to: 9pm)
-
-Farewell, UNIverse traveller. May our paths intersect again soon!
+Farewell, Universe traveller. May our paths intersect again soon!
